requirements.txt
streamlit==1.43.2  
streamlit-option-menu==0.3.6
python-dotenv==0.21.1
openai>=1.3.0              
tensorflow==2.15.0
keras==2.15.0             
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1            
pydub==0.25.1
soundfile==0.12.1
scikit-learn==1.3.0
nltk==3.8.1
numba==0.57
scipy==1.11.3
matplotlib==3.8.2
seaborn==0.13.2
opencv-python==4.7.0.72
rembg==2.0.50   
Pillow==9.3.0
bcrypt==4.0.1
numpy==1.23.5           
pytube==15.0.0          
requests==2.28.1         
streamlit-cookies-manager
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
ffmpeg-python==0.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gotrue==2.11.4
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
multidict==6.2.0
namex==0.0.8
narwhals==1.30.0
numba==0.61.0
numpy==2.1.3
openai==1.66.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
postgrest==0.19.3
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.1
qrcode==8.0
realtime==2.4.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
storage3==0.11.3
streamlit==1.44.1
streamlit-cookies-controller==0.0.4
streamlit-cookies-manager==0.2.0
streamlit-option-menu==0.4.0
StrEnum==0.4.15
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
streamlit-toggle
streamlit-toggle-switch
supabase==2.15.0
supafunc==0.9.3
supabase-py==1.0.3
supabase-py==1.0.3